livis_requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.1.5
django-cors-headers==3.4.0
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
numpy
olefile
opencv-python==4.3.0.36
packaging==21.0
pandas==1.3.4
passlib==1.7.4
Pillow==8.4.0
pycparser==2.20
pymongo==3.10.1
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2021.3
redis==2.10.6
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six
serial
serial-tool
sqlparse==0.4.2
typing-extensions
uritemplate==4.1.1
urllib3==1.26.7
xlsxwriter